requirements.txt
# TrialWatch - Core dependencies
requests>=2.31.0
pydantic>=2.4.0  
python-dotenv>=1.0.0
psycopg>=3.2.0
tenacity>=8.0.0

# EU CTIS Harvester dependencies
playwright>=1.40.0
pendulum>=3.0.0
boto3>=1.34.0

# Optional for development
pytest>=7.0.0
